+++ examples/mgard-x/AdaptiveResolution/SparseFlyingCells.hpp
@@ -253,9 +253,9 @@ public:
     tbz = 1;
     tby = 1;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
 
     return Task(functor, gridz, gridy, gridx, tbz, tby, tbx, sm_size,
                 queue_idx);
@@ -281,9 +281,9 @@ public:
     tbz = 1;
     tby = 1;
     tbx = F;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
 
     return Task(functor, gridz, gridy, gridx, tbz, tby, tbx, sm_size,
                 queue_idx);
