+++ include/mgard-x/DataRefactoring/InCacheBlock/Autocorrelation8x8x8.hpp
@@ -214,9 +214,9 @@ public:
     tbz = 8;
     tby = 8;
     tbx = 8;
-    gridz = ceil((float)total_thread_z / tbz);
-    gridy = ceil((float)total_thread_y / tby);
-    gridx = ceil((float)total_thread_x / tbx);
+    gridz = ceil((double)total_thread_z / tbz);
+    gridy = ceil((double)total_thread_y / tby);
+    gridx = ceil((double)total_thread_x / tbx);
 
     return Task(functor, gridz, gridy, gridx, tbz, tby, tbx, sm_size, queue_idx,
                 std::string(Name));
